arcboot initrd support patch
Hi Guido,

Attached is a proposed patch for initrd support in arcboot. It's
mostly based on an old patch submitted long ago, with some changes.

I'm using what should be the max page size for the machine to align
the initrd address, that's the offset of the first section of the ELF
image (found that in IRIX' elf manpage, thought it'd be nice to be
able to load kernels with different page sizes).

I've changed the order of the params[] array slightly to add the
initrd to the mix.

It looks like gcc 4.2 brings a ton of new warnings while building
arcboot :/

Tested on IP22 with a standard debian initrd.

JB.

--
 Julien BLACHE <jblache@debian.org>  |  Debian, because code matters more
 Debian & GNU/Linux Developer        |       <http://www.debian.org>
 Public key available on <http://www.jblache.org> - KeyID: F5D6 5169
 GPG Fingerprint : 935A 79F1 C8B3 3521 FD62 7CC7 CD61 4FD7 F5D6 5169

Signed-off-by: Guido Guenther <agx@sigxcpu.org>