apply an IP22 tftploader hack so the kernel+d-i initrd fit into the reserved space
http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=452798

Tip22 finds this memory map on IP22:

1) Free Memory(3) segment found at (0x88002000,0x88740000)
2) Free Memory(3) segment found at (0x88800000,0x89002000)
3) Free Memory(3) segment found at (0x896f2000,0x94000000)

(1) is where the running kernel + initrd goes. (2) is where we tell the
prom to download the tftpboot image to and (3) is lots of free space.
The 64bit kernel (7MB) is just so much bigger than the 32bit one.

This hack puts the tftp boot file into (3) and the initrd into (2) -
it succesfully boots a kernel with an initramfs on my R5K Indy. I'll
have to change some of the reservation code so we don't waste so much
ram before I can fix this for real though.